+++ datalad/cmd.py
@@ -6,65 +6,84 @@
 #   copyright and license terms.
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
+"""
+Wrapper for command calls, allowing for dry runs and output handling
 
-import os
-import commands
+"""
 
+
+import subprocess
+import sys
 import logging
-lgr = logging.getLogger('datalad.files')
 
-def dry(s, dry_):
-    """Small helper for dry runs
-    """
-    if dry_:
-        return "DRY " + s
-    return s
+lgr = logging.getLogger('datalad.cmd')
+
 
 class Runner(object):
-    """Helper to run commands and functions while allowing for dry_run and later reporting
+    """Provides a wrapper for system calls.
+
+    An object of this class provides a method to make system calls.
+    Outputs (stdout and stderr) are streamed to stdout during execution as if you were calling it from command line.
+    Additionally allows for dry runs. This is achieved by initializing the `Runner` with `dry=True`.
+    The Runner will then collect all calls as strings in `commands`.
     """
 
     __slots__ = ['commands', 'dry']
 
     def __init__(self, dry=False):
+        self.dry = dry
         self.commands = []
-        self.dry = dry                    # TODO: make use of it
 
-    # TODO -- remove dry_run specification here -- use constructor parameter
-    #def __call__(self, cmd, dry_run=False):
-    def getstatusoutput(self, cmd, dry_run=None):
-        """A wrapper around commands.getstatusoutput
+    def run(self, cmd):
+        """Runs the command `cmd` using shell.
 
-        Provides improved logging for debugging purposes and raises
-        RuntimeError exception in case of non-0 return
+        `cmd` is called and uses stdout whereas stderr is captured and logged.
+        In case of dry-mode `cmd` is just added to `commands`.
+
+        Parameters
+        ----------
+        cmd : str
+          String defining the command call.
+
+        Returns
+        -------
+        Status code as returned by the called command or `None` in case of dry-mode.
         """
-        self.log("Running: %s" % (cmd,))
+
+        self.log("Running: %s" % cmd)
+
         if not self.dry:
-            # status, output = commands.getstatusoutput(cmd)
-            # doing manually for improved debugging
-            pipe = os.popen('{ ' + cmd + '; } 2>&1', 'r')
-            output = ''
-            for line in iter(pipe.readline, ''):
-                self.log("| " + line.rstrip('\n'), level=5)
-                output += line
-            if output[-1:] == '\n': output = output[:-1]
-            status = pipe.close() or 0
+
+            proc = subprocess.Popen(cmd, stdout=sys.stdout, stderr=subprocess.PIPE, shell=True)
+            # shell=True allows for piping, multiple commands, etc., but that implies to not use shlex.split()
+            # and is considered to be a security hazard. So be careful with input.
+            # Alternatively we would have to parse `cmd` and create multiple subprocesses.
+
+            while proc.poll() is None:
+                err_line = proc.stderr.readline()
+                if err_line != '':
+                    self.log("stderr: %s" % err_line)
+                    #TODO: log per line or as a whole?
+
+            status = proc.poll()
 
             if not status in [0, None]:
-                msg = "Failed to run %r. Exit code=%d output=%s" \
-                      % (cmd, status, output)
+                msg = "Failed to run %r. Exit code=%d" % (cmd, status)
                 lgr.error(msg)
                 raise RuntimeError(msg)
+
             else:
-                self.log("Finished running %r with status %s" % (cmd, status),
-                         level=8)
-                return status, output
+                self.log("Finished running %r with status %s" % (cmd, status), level=8)
+                return status
+
         else:
             self.commands.append(cmd)
-        return None, None
+        return None
 
     def drycall(self, f, *args, **kwargs):
         """Helper to unify collection of logging all "dry" actions.
+
+        Calls `f` if `Runner`-object is not in dry-mode. Adds `f` along with its arguments to `commands` otherwise.
 
         f : callable
         *args, **kwargs:
@@ -73,20 +92,19 @@ class Runner(object):
         if self.dry:
             self.commands.append("%s args=%s kwargs=%s" % (f, args, kwargs))
         else:
-            f(*args, **kwargs)
+            return f(*args, **kwargs)
 
-    def log(self, msg, level=None):
-        if level is None:
-            logf = lgr.debug
-        else:
-            logf = lambda msg: lgr.log(level, msg)
+    def log(self, msg, level=logging.DEBUG):
         if self.dry:
-            lgr.debug("DRY: %s" % msg)
+            lgr.log(level, "DRY: %s" % msg)
         else:
-            lgr.debug(msg)
+            lgr.log(level, msg)
 
-#getstatusoutput = Runner()
 
+# ####
+# Preserve from previous version
+# TODO: document intention
+# ####
 # this one might get under Runner for better output/control
 def link_file_load(src, dst, dry_run=False):
     """Just a little helper to hardlink files's load
