+++ datalad/tests/test_cmd.py
@@ -14,7 +14,7 @@ import sys
 
 from nose.tools import assert_is, assert_equal, assert_false, assert_true, assert_greater
 
-from datalad.cmd2 import Runner
+from datalad.cmd import Runner
 from datalad.tests.utils import with_tempfile, assert_cwd_unchanged
 
 
@@ -28,15 +28,15 @@ def test_runner_dry(tempfile):
     cmd = 'echo Testing dry run > %s' % tempfile
     ret = runner.run(cmd)
     assert_is(None, ret, "Dry run of: %s resulted in exitcode %s" % (cmd, ret))
-    assert_equal(runner.cmdBuffer.__str__(), ("['%s']" % cmd),
-                 "Dry run of: %s resulted in buffer: %s" % (cmd, runner.cmdBuffer.__str__()))
+    assert_equal(runner.commands.__str__(), ("['%s']" % cmd),
+                 "Dry run of: %s resulted in buffer: %s" % (cmd, runner.commands.__str__()))
     assert_false(os.path.exists(tempfile))
 
     # test dry python function call
     output = runner.drycall(os.path.join, 'foo', 'bar')
     assert_is(None, output, "Drycall of: os.path.join, 'foo', 'bar' returned %s" % output)
-    assert_greater(runner.cmdBuffer.__str__().find('join'), -1,
-                   "Drycall of: os.path.join, 'foo', 'bar' resulted in buffer: %s" % runner.cmdBuffer.__str__())
+    assert_greater(runner.commands.__str__().find('join'), -1,
+                   "Drycall of: os.path.join, 'foo', 'bar' resulted in buffer: %s" % runner.commands.__str__())
 
 
 @assert_cwd_unchanged
@@ -48,15 +48,15 @@ def test_runner(tempfile):
     cmd = 'echo Testing real run > %s' % tempfile
     ret = runner.run(cmd)
     assert_equal(0, ret, "Run of: %s resulted in exitcode %s" % (cmd, ret))
-    assert_equal(runner.cmdBuffer, [], "Run of: %s resulted in non-empty buffer: %s" % (cmd, runner.cmdBuffer.__str__()))
+    assert_equal(runner.commands, [], "Run of: %s resulted in non-empty buffer: %s" % (cmd, runner.commands.__str__()))
     assert_true(os.path.exists(tempfile), "Run of: %s resulted with non-existing file %s" % (cmd, tempfile))
 
     # test non-dry python function call
     output = runner.drycall(os.path.join, 'foo', 'bar')
     assert_equal(os.path.join('foo', 'bar'), output,
                  "Drycall of: os.path.join, 'foo', 'bar' returned %s" % output)
-    assert_equal(runner.cmdBuffer.__str__().find('os.path.join'), -1,
-                   "Drycall of: os.path.join, 'foo', 'bar' resulted in buffer: %s" % runner.cmdBuffer.__str__())
+    assert_equal(runner.commands.__str__().find('os.path.join'), -1,
+                   "Drycall of: os.path.join, 'foo', 'bar' resulted in buffer: %s" % runner.commands.__str__())
 
 
 def test_runner_log_stderr():
@@ -66,7 +66,7 @@ def test_runner_log_stderr():
     cmd = 'echo stderr-Message should be logged >&2'
     ret = runner.run(cmd)
     assert_equal(0, ret, "Run of: %s resulted in exitcode %s" % (cmd, ret))
-    assert_equal(runner.cmdBuffer, [], "Run of: %s resulted in non-empty buffer: %s" % (cmd, runner.cmdBuffer.__str__()))
+    assert_equal(runner.commands, [], "Run of: %s resulted in non-empty buffer: %s" % (cmd, runner.commands.__str__()))
 
 
 def test_runner_heavy_output():
